pyproject.toml
[project]
name = "nlp-to-sql-app"
version = "0.1.0"
description = "NLP to SQL Query Application using LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "streamlit>=1.28.0",
    "pandas>=1.5.0"
]
